requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
bidict==0.21.2
black==21.5b1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coverage==5.3
decorator==4.4.2
discord.py==1.5.1
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
jedi==0.17.2
more-itertools==8.6.0
multidict==4.7.6
mypy==0.812
mypy-extensions==0.4.3
numpy==1.19.4
packaging==20.4
parso==0.7.1
pathspec==0.8.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.2
python-dotenv==0.15.0
regex==2020.11.13
rope==0.18.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
validators==0.18.2
wcwidth==0.2.5
websockets==8.1
yapf==0.31.0
yarl==1.6.3
youtube-dl==2021.5.16
zipp==3.4.0
berserk==0.10.0
